pyproject.toml
[project]
name = "structer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "drf-spectacular-sidecar>=2026.1.1",
    "pillow>=12.1.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "flake8>=7.3.0",
    "isort>=7.0.0",
    "pre-commit>=4.5.1",
]
